requirements.txt
paddlepaddle-gpu>=2.5.0
paddleocr>=2.7.0
albumentations>=1.3.0
symspellpy>=6.7.7
pyctcdecode>=0.5.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
onnxruntime-gpu>=1.15.0
transformers>=4.30.0
editdistance>=0.6.3
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
trdg>=1.8.0
paddle2onnx>=1.0.0
torch>=2.0.0
